improved contract naming format, fix tests

+++ contracts/test/RouterEventEmitter.sol
@@ -4,95 +4,141 @@ pragma solidity =0.7.6;
 import '../interfaces/IImpossibleRouter01.sol';
 
 contract RouterEventEmitter {
-    event Amounts(uint[] amounts);
+    event Amounts(uint256[] amounts);
 
     receive() external payable {}
 
     function swapExactTokensForTokens(
         address router,
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapExactTokensForTokens.selector, amountIn, amountOutMin, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapExactTokensForTokens.selector,
+                    amountIn,
+                    amountOutMin,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 
     function swapTokensForExactTokens(
         address router,
-        uint amountOut,
-        uint amountInMax,
+        uint256 amountOut,
+        uint256 amountInMax,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapTokensForExactTokens.selector, amountOut, amountInMax, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapTokensForExactTokens.selector,
+                    amountOut,
+                    amountInMax,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 
     function swapExactETHForTokens(
         address router,
-        uint amountOutMin,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external payable {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapExactETHForTokens.selector, amountOutMin, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapExactETHForTokens.selector,
+                    amountOutMin,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 
     function swapTokensForExactETH(
         address router,
-        uint amountOut,
-        uint amountInMax,
+        uint256 amountOut,
+        uint256 amountInMax,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapTokensForExactETH.selector, amountOut, amountInMax, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapTokensForExactETH.selector,
+                    amountOut,
+                    amountInMax,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 
     function swapExactTokensForETH(
         address router,
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapExactTokensForETH.selector, amountIn, amountOutMin, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapExactTokensForETH.selector,
+                    amountIn,
+                    amountOutMin,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 
     function swapETHForExactTokens(
         address router,
-        uint amountOut,
+        uint256 amountOut,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external payable {
-        (bool success, bytes memory returnData) = router.call(abi.encodeWithSelector(
-            IImpossibleRouter01(router).swapETHForExactTokens.selector, amountOut, path, to, deadline
-        ));
+        (bool success, bytes memory returnData) =
+            router.call(
+                abi.encodeWithSelector(
+                    IImpossibleRouter01(router).swapETHForExactTokens.selector,
+                    amountOut,
+                    path,
+                    to,
+                    deadline
+                )
+            );
         assert(success);
-        emit Amounts(abi.decode(returnData, (uint[])));
+        emit Amounts(abi.decode(returnData, (uint256[])));
     }
 }
